% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidejoin-package.R
\docType{package}
\name{slidejoin-package}
\alias{slidejoin}
\alias{slidejoin-package}
\title{slidejoin: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/slidejoin}
  \item Report bugs at \url{https://github.com/DavisVaughan/slidejoin/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
